Breaking up GL agnostic code from 4.1 specific code
