+++ libraries/gpu-gl/src/gpu/gl/GLBackend.h
@@ -8,8 +8,8 @@
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
-#ifndef hifi_gpu_GLBackend_h
-#define hifi_gpu_GLBackend_h
+#ifndef hifi_gpu_gl_GLBackend_h
+#define hifi_gpu_gl_GLBackend_h
 
 #include <assert.h>
 #include <functional>
@@ -26,325 +26,34 @@
 #include <gpu/Forward.h>
 #include <gpu/Context.h>
 
-
-#define GPU_CORE 1
-#define GPU_LEGACY 0
-#define GPU_CORE_41 410
-#define GPU_CORE_43 430
-#define GPU_CORE_MINIMUM GPU_CORE_41
-#define GPU_FEATURE_PROFILE GPU_CORE
-
-#if defined(__APPLE__)
-
-#define GPU_INPUT_PROFILE GPU_CORE_41
-
-#else
-
-#define GPU_INPUT_PROFILE GPU_CORE_43
-
-#endif
-
-
-#if (GPU_INPUT_PROFILE == GPU_CORE_43)
-// Deactivate SSBO for now, we've run into some issues
-// on GL 4.3 capable GPUs not behaving as expected
-//#define GPU_SSBO_DRAW_CALL_INFO
-#endif
+#include "GLShared.h"
 
 namespace gpu { namespace gl {
 
-class GLTextureTransferHelper;
-
 class GLBackend : public Backend {
-
     // Context Backend static interface required
     friend class gpu::Context;
     static void init();
     static Backend* createBackend();
-    static bool makeProgram(Shader& shader, const Shader::BindingSet& bindings);
+    static bool makeProgram(Shader& shader, const Shader::BindingSet& slotBindings);
 
+protected:
     explicit GLBackend(bool syncCache);
     GLBackend();
 public:
-    static Context::Size getDedicatedMemory();
+    ~GLBackend();
 
-    virtual ~GLBackend();
-
-    virtual void render(Batch& batch);
+    void render(Batch& batch) final;
 
     // This call synchronize the Full Backend cache with the current GLState
     // THis is only intended to be used when mixing raw gl calls with the gpu api usage in order to sync
     // the gpu::Backend state with the true gl state which has probably been messed up by these ugly naked gl calls
     // Let's try to avoid to do that as much as possible!
-    virtual void syncCache();
+    void syncCache() final;
 
     // This is the ugly "download the pixels to sysmem for taking a snapshot"
     // Just avoid using it, it's ugly and will break performances
-    virtual void downloadFramebuffer(const FramebufferPointer& srcFramebuffer, const Vec4i& region, QImage& destImage);
-
-    static void checkGLStackStable(std::function<void()> f);
-
-    
-
-    class GLBuffer : public GPUObject {
-    public:
-        const GLuint _buffer;
-        const GLuint _size;
-        const Stamp _stamp;
-
-        GLBuffer(const Buffer& buffer, GLBuffer* original = nullptr);
-        ~GLBuffer();
-
-        virtual void transfer();
-
-    private:
-        bool getNextTransferBlock(GLintptr& outOffset, GLsizeiptr& outSize, size_t& currentPage) const;
-            
-        // The owning texture
-        const Buffer& _gpuBuffer;
-    };
-
-    static GLBuffer* syncGPUObject(const Buffer& buffer);
-    static GLuint getBufferID(const Buffer& buffer);
-
-    class GLTexture : public GPUObject {
-    public:
-        // The public gl texture object
-        GLuint _texture{ 0 };
-
-        const Stamp _storageStamp;
-        Stamp _contentStamp { 0 };
-        const GLenum _target;
-        const uint16 _maxMip;
-        const uint16 _minMip;
-        const bool _transferrable;
-
-        struct DownsampleSource {
-            using Pointer = std::shared_ptr<DownsampleSource>;
-            DownsampleSource(GLTexture& oldTexture);
-            ~DownsampleSource();
-            const GLuint _texture;
-            const uint16 _minMip;
-            const uint16 _maxMip;
-        };
-
-        DownsampleSource::Pointer _downsampleSource;
-
-        GLTexture(bool transferrable, const gpu::Texture& gpuTexture);
-        GLTexture(GLTexture& originalTexture, const gpu::Texture& gpuTexture);
-        ~GLTexture();
-
-        // Return a floating point value indicating how much of the allowed 
-        // texture memory we are currently consuming.  A value of 0 indicates 
-        // no texture memory usage, while a value of 1 indicates all available / allowed memory
-        // is consumed.  A value above 1 indicates that there is a problem.
-        static float getMemoryPressure();
-
-        void withPreservedTexture(std::function<void()> f);
-
-        void createTexture();
-        void allocateStorage();
-
-        GLuint size() const { return _size; }
-        GLuint virtualSize() const { return _virtualSize; }
-
-        void updateSize();
-
-        enum SyncState {
-            // The texture is currently undergoing no processing, although it's content
-            // may be out of date, or it's storage may be invalid relative to the 
-            // owning GPU texture
-            Idle,
-            // The texture has been queued for transfer to the GPU
-            Pending,
-            // The texture has been transferred to the GPU, but is awaiting
-            // any post transfer operations that may need to occur on the 
-            // primary rendering thread
-            Transferred,
-        };
-
-        void setSyncState(SyncState syncState) { _syncState = syncState; }
-        SyncState getSyncState() const { return _syncState; }
-
-        // Is the storage out of date relative to the gpu texture?
-        bool isInvalid() const;
-
-        // Is the content out of date relative to the gpu texture?
-        bool isOutdated() const;
-
-        // Is the texture in a state where it can be rendered with no work?
-        bool isReady() const;
-
-        // Is this texture pushing us over the memory limit?
-        bool isOverMaxMemory() const;
-
-        // Move the image bits from the CPU to the GPU
-        void transfer() const;
-
-        // Execute any post-move operations that must occur only on the main thread
-        void postTransfer();
-
-        uint16 usedMipLevels() const { return (_maxMip - _minMip) + 1; }
-
-        static const size_t CUBE_NUM_FACES = 6;
-        static const GLenum CUBE_FACE_LAYOUT[6];
-        
-    private:
-        friend class GLTextureTransferHelper;
-
-        GLTexture(bool transferrable, const gpu::Texture& gpuTexture, bool init);
-        // at creation the true texture is created in GL
-        // it becomes public only when ready.
-        GLuint _privateTexture{ 0 };
-
-        const std::vector<GLenum>& getFaceTargets() const;
-
-        void setSize(GLuint size);
-
-        const GLuint _virtualSize; // theorical size as expected
-        GLuint _size { 0 }; // true size as reported by the gl api
-
-        void transferMip(uint16_t mipLevel, uint8_t face = 0) const;
-
-        // The owning texture
-        const Texture& _gpuTexture;
-        std::atomic<SyncState> _syncState { SyncState::Idle };
-    };
-    static GLTexture* syncGPUObject(const TexturePointer& texture, bool needTransfer = true);
-    static GLuint getTextureID(const TexturePointer& texture, bool sync = true);
-
-    // very specific for now
-    static void syncSampler(const Sampler& sampler, Texture::Type type, const GLTexture* object);
-
-    class GLShader : public GPUObject {
-    public:
-        enum Version {
-            Mono = 0,
-
-            NumVersions
-        };
-
-        struct ShaderObject {
-            GLuint glshader{ 0 };
-            GLuint glprogram{ 0 };
-            GLint transformCameraSlot{ -1 };
-            GLint transformObjectSlot{ -1 };
-        };
-
-        using ShaderObjects = std::array< ShaderObject, NumVersions >;
-
-        using UniformMapping = std::map<GLint, GLint>;
-        using UniformMappingVersions = std::vector<UniformMapping>;
-        
-        GLShader();
-        ~GLShader();
-
-        ShaderObjects _shaderObjects;
-        UniformMappingVersions _uniformMappings;
-
-        GLuint getProgram(Version version = Mono) const {
-            return _shaderObjects[version].glprogram;
-        }
-
-        GLint getUniformLocation(GLint srcLoc, Version version = Mono) {
-            // THIS will be used in the future PR as we grow the number of versions
-            // return _uniformMappings[version][srcLoc];
-            return srcLoc;
-        }
-
-    };
-    static GLShader* syncGPUObject(const Shader& shader);
-
-    class GLState : public GPUObject {
-    public:
-        class Command {
-        public:
-            virtual void run(GLBackend* backend) = 0;
-            Command() {}
-            virtual ~Command() {};
-        };
-
-        template <class T> class Command1 : public Command {
-        public:
-            typedef void (GLBackend::*GLFunction)(T);
-            void run(GLBackend* backend) { (backend->*(_func))(_param); }
-            Command1(GLFunction func, T param) : _func(func), _param(param) {};
-            GLFunction _func;
-            T _param;
-        };
-        template <class T, class U> class Command2 : public Command {
-        public:
-            typedef void (GLBackend::*GLFunction)(T, U);
-            void run(GLBackend* backend) { (backend->*(_func))(_param0, _param1); }
-            Command2(GLFunction func, T param0, U param1) : _func(func), _param0(param0), _param1(param1) {};
-            GLFunction _func;
-            T _param0;
-            U _param1;
-        };
-
-        template <class T, class U, class V> class Command3 : public Command {
-        public:
-            typedef void (GLBackend::*GLFunction)(T, U, V);
-            void run(GLBackend* backend) { (backend->*(_func))(_param0, _param1, _param2); }
-            Command3(GLFunction func, T param0, U param1, V param2) : _func(func), _param0(param0), _param1(param1), _param2(param2) {};
-            GLFunction _func;
-            T _param0;
-            U _param1;
-            V _param2;
-        };
-
-        typedef std::shared_ptr< Command > CommandPointer;
-        typedef std::vector< CommandPointer > Commands;
-
-        Commands _commands;
-        Stamp _stamp;
-        State::Signature _signature;
-
-        GLState();
-        ~GLState();
-
-        // The state commands to reset to default,
-        static const Commands _resetStateCommands;
-
-        friend class GLBackend;
-    };
-    static GLState* syncGPUObject(const State& state);
-
-    class GLPipeline : public GPUObject {
-    public:
-        GLShader* _program = 0;
-        GLState* _state = 0;
-
-        GLPipeline();
-        ~GLPipeline();
-    };
-    static GLPipeline* syncGPUObject(const Pipeline& pipeline);
-
-
-    class GLFramebuffer : public GPUObject {
-    public:
-        GLuint _fbo = 0;
-        std::vector<GLenum> _colorBuffers;
-        Stamp _depthStamp { 0 };
-        std::vector<Stamp> _colorStamps;
-
-
-        GLFramebuffer();
-        ~GLFramebuffer();
-    };
-    static GLFramebuffer* syncGPUObject(const Framebuffer& framebuffer);
-    static GLuint getFramebufferID(const FramebufferPointer& framebuffer);
-
-    class GLQuery : public GPUObject {
-    public:
-        GLuint _qo = 0;
-        GLuint64 _result = 0;
-
-        GLQuery();
-        ~GLQuery();
-    };
-    static GLQuery* syncGPUObject(const Query& query);
-    static GLuint getQueryID(const QueryPointer& query);
+    virtual void downloadFramebuffer(const FramebufferPointer& srcFramebuffer, const Vec4i& region, QImage& destImage) final;
 
 
     static const int MAX_NUM_ATTRIBUTES = Stream::NUM_INPUT_SLOTS;
@@ -362,70 +71,131 @@ public:
     static const int MAX_NUM_RESOURCE_TEXTURES = 16;
     size_t getMaxNumResourceTextures() const { return MAX_NUM_RESOURCE_TEXTURES; }
 
+    // Draw Stage
+    virtual void do_draw(Batch& batch, size_t paramOffset) = 0;
+    virtual void do_drawIndexed(Batch& batch, size_t paramOffset) = 0;
+    virtual void do_drawInstanced(Batch& batch, size_t paramOffset) = 0;
+    virtual void do_drawIndexedInstanced(Batch& batch, size_t paramOffset) = 0;
+    virtual void do_multiDrawIndirect(Batch& batch, size_t paramOffset) = 0;
+    virtual void do_multiDrawIndexedIndirect(Batch& batch, size_t paramOffset) = 0;
+
+    // Input Stage
+    virtual void do_setInputFormat(Batch& batch, size_t paramOffset) final;
+    virtual void do_setInputBuffer(Batch& batch, size_t paramOffset) final;
+    virtual void do_setIndexBuffer(Batch& batch, size_t paramOffset) final;
+    virtual void do_setIndirectBuffer(Batch& batch, size_t paramOffset) final;
+    virtual void do_generateTextureMips(Batch& batch, size_t paramOffset) final;
+
+    // Transform Stage
+    virtual void do_setModelTransform(Batch& batch, size_t paramOffset) final;
+    virtual void do_setViewTransform(Batch& batch, size_t paramOffset) final;
+    virtual void do_setProjectionTransform(Batch& batch, size_t paramOffset) final;
+    virtual void do_setViewportTransform(Batch& batch, size_t paramOffset) final;
+    virtual void do_setDepthRangeTransform(Batch& batch, size_t paramOffset) final;
+
+    // Uniform Stage
+    virtual void do_setUniformBuffer(Batch& batch, size_t paramOffset) final;
+
+    // Resource Stage
+    virtual void do_setResourceTexture(Batch& batch, size_t paramOffset) final;
+
+    // Pipeline Stage
+    virtual void do_setPipeline(Batch& batch, size_t paramOffset) final;
+
+    // Output stage
+    virtual void do_setFramebuffer(Batch& batch, size_t paramOffset) final;
+    virtual void do_clearFramebuffer(Batch& batch, size_t paramOffset) final;
+    virtual void do_blit(Batch& batch, size_t paramOffset) = 0;
+
+    // Query section
+    virtual void do_beginQuery(Batch& batch, size_t paramOffset) final;
+    virtual void do_endQuery(Batch& batch, size_t paramOffset) final;
+    virtual void do_getQuery(Batch& batch, size_t paramOffset) final;
+
+    // Reset stages
+    virtual void do_resetStages(Batch& batch, size_t paramOffset) final;
+
+    virtual void do_runLambda(Batch& batch, size_t paramOffset) final;
+
+    virtual void do_startNamedCall(Batch& batch, size_t paramOffset) final;
+    virtual void do_stopNamedCall(Batch& batch, size_t paramOffset) final;
+
+    virtual void do_pushProfileRange(Batch& batch, size_t paramOffset) final;
+    virtual void do_popProfileRange(Batch& batch, size_t paramOffset) final;
+
+    // TODO: As long as we have gl calls explicitely issued from interface
+    // code, we need to be able to record and batch these calls. THe long 
+    // term strategy is to get rid of any GL calls in favor of the HIFI GPU API
+    virtual void do_glActiveBindTexture(Batch& batch, size_t paramOffset) final;
+
+    virtual void do_glUniform1i(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform1f(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform2f(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform3f(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform4f(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform3fv(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform4fv(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniform4iv(Batch& batch, size_t paramOffset) final;
+    virtual void do_glUniformMatrix4fv(Batch& batch, size_t paramOffset) final;
+
+    virtual void do_glColor4f(Batch& batch, size_t paramOffset) final;
+
     // The State setters called by the GLState::Commands when a new state is assigned
-    void do_setStateFillMode(int32 mode);
-    void do_setStateCullMode(int32 mode);
-    void do_setStateFrontFaceClockwise(bool isClockwise);
-    void do_setStateDepthClampEnable(bool enable);
-    void do_setStateScissorEnable(bool enable);
-    void do_setStateMultisampleEnable(bool enable);
-    void do_setStateAntialiasedLineEnable(bool enable);
+    virtual void do_setStateFillMode(int32 mode) final;
+    virtual void do_setStateCullMode(int32 mode) final;
+    virtual void do_setStateFrontFaceClockwise(bool isClockwise) final;
+    virtual void do_setStateDepthClampEnable(bool enable) final;
+    virtual void do_setStateScissorEnable(bool enable) final;
+    virtual void do_setStateMultisampleEnable(bool enable) final;
+    virtual void do_setStateAntialiasedLineEnable(bool enable) final;
+    virtual void do_setStateDepthBias(Vec2 bias) final;
+    virtual void do_setStateDepthTest(State::DepthTest test) final;
+    virtual void do_setStateStencil(State::StencilActivation activation, State::StencilTest frontTest, State::StencilTest backTest) final;
+    virtual void do_setStateAlphaToCoverageEnable(bool enable) final;
+    virtual void do_setStateSampleMask(uint32 mask) final;
+    virtual void do_setStateBlend(State::BlendFunction blendFunction) final;
+    virtual void do_setStateColorWriteMask(uint32 mask) final;
+    virtual void do_setStateBlendFactor(Batch& batch, size_t paramOffset) final;
+    virtual void do_setStateScissorRect(Batch& batch, size_t paramOffset) final;
 
-    void do_setStateDepthBias(Vec2 bias);
-    void do_setStateDepthTest(State::DepthTest test);
-
-    void do_setStateStencil(State::StencilActivation activation, State::StencilTest frontTest, State::StencilTest backTest);
-
-    void do_setStateAlphaToCoverageEnable(bool enable);
-    void do_setStateSampleMask(uint32 mask);
-
-    void do_setStateBlend(State::BlendFunction blendFunction);
-
-    void do_setStateColorWriteMask(uint32 mask);
-    
 protected:
-    static const size_t INVALID_OFFSET = (size_t)-1;
 
-    bool _inRenderTransferPass;
+    virtual GLuint getFramebufferID(const FramebufferPointer& framebuffer) = 0;
+    virtual GLFramebuffer* syncGPUObject(const Framebuffer& framebuffer) = 0;
+
+    virtual GLuint getBufferID(const Buffer& buffer) = 0;
+    virtual GLBuffer* syncGPUObject(const Buffer& buffer) = 0;
+
+    virtual GLuint getTextureID(const TexturePointer& texture, bool needTransfer = true) = 0;
+    virtual GLTexture* syncGPUObject(const TexturePointer& texture, bool sync = true) = 0;
+
+    virtual GLuint getQueryID(const QueryPointer& query) = 0;
+    virtual GLQuery* syncGPUObject(const Query& query) = 0;
+
+    static const size_t INVALID_OFFSET = (size_t)-1;
+    bool _inRenderTransferPass { false };
+    int32_t _uboAlignment { 0 };
+    int _currentDraw { -1 };
 
     void renderPassTransfer(Batch& batch);
     void renderPassDraw(Batch& batch);
-
     void setupStereoSide(int side);
 
-    void initTextureTransferHelper();
-    static void transferGPUObject(const TexturePointer& texture);
+    virtual void initInput() final;
+    virtual void killInput() final;
+    virtual void syncInputStateCache() final;
+    virtual void resetInputStage() final;
+    virtual void updateInput() = 0;
 
-    static std::shared_ptr<GLTextureTransferHelper> _textureTransferHelper;
-
-    // Draw Stage
-    void do_draw(Batch& batch, size_t paramOffset);
-    void do_drawIndexed(Batch& batch, size_t paramOffset);
-    void do_drawInstanced(Batch& batch, size_t paramOffset);
-    void do_drawIndexedInstanced(Batch& batch, size_t paramOffset);
-    void do_multiDrawIndirect(Batch& batch, size_t paramOffset);
-    void do_multiDrawIndexedIndirect(Batch& batch, size_t paramOffset);
-    
-    // Input Stage
-    void do_setInputFormat(Batch& batch, size_t paramOffset);
-    void do_setInputBuffer(Batch& batch, size_t paramOffset);
-    void do_setIndexBuffer(Batch& batch, size_t paramOffset);
-    void do_setIndirectBuffer(Batch& batch, size_t paramOffset);
-
-    void initInput();
-    void killInput();
-    void syncInputStateCache();
-    void updateInput();
-    void resetInputStage();
     struct InputStageState {
-        bool _invalidFormat = true;
+        bool _invalidFormat { true };
         Stream::FormatPointer _format;
 
         typedef std::bitset<MAX_NUM_ATTRIBUTES> ActivationCache;
-        ActivationCache _attributeActivation;
+        ActivationCache _attributeActivation { 0 };
 
         typedef std::bitset<MAX_NUM_INPUT_BUFFERS> BuffersState;
-        BuffersState _invalidBuffers;
+        BuffersState _invalidBuffers { 0 };
 
         Buffers _buffers;
         Offsets _bufferOffsets;
@@ -435,39 +205,23 @@ protected:
         glm::vec4 _colorAttribute{ 0.0f };
 
         BufferPointer _indexBuffer;
-        Offset _indexBufferOffset;
-        Type _indexBufferType;
+        Offset _indexBufferOffset { 0 };
+        Type _indexBufferType { UINT32 };
         
         BufferPointer _indirectBuffer;
         Offset _indirectBufferOffset{ 0 };
         Offset _indirectBufferStride{ 0 };
 
-        GLuint _defaultVAO;
+        GLuint _defaultVAO { 0 };
 
         InputStageState() :
-            _invalidFormat(true),
-            _format(0),
-            _attributeActivation(0),
-            _invalidBuffers(0),
-            _buffers(_invalidBuffers.size(), BufferPointer(0)),
+            _buffers(_invalidBuffers.size()),
             _bufferOffsets(_invalidBuffers.size(), 0),
             _bufferStrides(_invalidBuffers.size(), 0),
-            _bufferVBOs(_invalidBuffers.size(), 0),
-            _indexBuffer(0),
-            _indexBufferOffset(0),
-            _indexBufferType(UINT32),
-            _defaultVAO(0)
-             {}
+            _bufferVBOs(_invalidBuffers.size(), 0) {}
     } _input;
 
-    // Transform Stage
-    void do_setModelTransform(Batch& batch, size_t paramOffset);
-    void do_setViewTransform(Batch& batch, size_t paramOffset);
-    void do_setProjectionTransform(Batch& batch, size_t paramOffset);
-    void do_setViewportTransform(Batch& batch, size_t paramOffset);
-    void do_setDepthRangeTransform(Batch& batch, size_t paramOffset);
-
-    void initTransform();
+    virtual void initTransform() = 0;
     void killTransform();
     // Synchronize the state cache of this Backend with the actual real state of the GL Context
     void syncTransformStateCache();
@@ -505,153 +259,74 @@ protected:
         void preUpdate(size_t commandIndex, const StereoState& stereo);
         void update(size_t commandIndex, const StereoState& stereo) const;
         void bindCurrentCamera(int stereoSide) const;
-        void transfer(const Batch& batch) const;
     } _transform;
 
-    int32_t _uboAlignment{ 0 };
+    virtual void transferTransformState(const Batch& batch) const = 0;
 
-
-    // Uniform Stage
-    void do_setUniformBuffer(Batch& batch, size_t paramOffset);
+    struct UniformStageState {
+        std::array<BufferPointer, MAX_NUM_UNIFORM_BUFFERS> _buffers;
+        //Buffers _buffers {  };
+    } _uniform;
 
     void releaseUniformBuffer(uint32_t slot);
     void resetUniformStage();
-    struct UniformStageState {
-        Buffers _buffers;
-
-        UniformStageState():
-            _buffers(MAX_NUM_UNIFORM_BUFFERS, nullptr)
-        {}
-    } _uniform;
     
-    // Resource Stage
-    void do_setResourceTexture(Batch& batch, size_t paramOffset);
-
     // update resource cache and do the gl unbind call with the current gpu::Texture cached at slot s
     void releaseResourceTexture(uint32_t slot);
 
     void resetResourceStage();
+
     struct ResourceStageState {
-        Textures _textures;
-
+        std::array<TexturePointer, MAX_NUM_RESOURCE_TEXTURES> _textures;
+        //Textures _textures { { MAX_NUM_RESOURCE_TEXTURES } };
         int findEmptyTextureSlot() const;
-
-        ResourceStageState():
-            _textures(MAX_NUM_RESOURCE_TEXTURES, nullptr)
-        {}
-
     } _resource;
+
     size_t _commandIndex{ 0 };
 
-    // Pipeline Stage
-    void do_setPipeline(Batch& batch, size_t paramOffset);
-    void do_setStateBlendFactor(Batch& batch, size_t paramOffset);
-    void do_setStateScissorRect(Batch& batch, size_t paramOffset);
-    
     // Standard update pipeline check that the current Program and current State or good to go for a
     void updatePipeline();
     // Force to reset all the state fields indicated by the 'toBeReset" signature
     void resetPipelineState(State::Signature toBeReset);
     // Synchronize the state cache of this Backend with the actual real state of the GL Context
     void syncPipelineStateCache();
-    // Grab the actual gl state into it's gpu::State equivalent. THis is used by the above call syncPipleineStateCache() 
-    void getCurrentGLState(State::Data& state);
     void resetPipelineStage();
 
     struct PipelineStageState {
-
         PipelinePointer _pipeline;
 
-        GLuint _program;
-        GLShader* _programShader;
-        bool _invalidProgram;
+        GLuint _program { 0 };
+        GLShader* _programShader { nullptr };
+        bool _invalidProgram { false };
 
-        State::Data _stateCache;
-        State::Signature _stateSignatureCache;
+        State::Data _stateCache { State::DEFAULT };
+        State::Signature _stateSignatureCache { 0 };
 
-        GLState* _state;
-        bool _invalidState = false;
-
-        PipelineStageState() :
-            _pipeline(),
-            _program(0),
-            _programShader(nullptr),
-            _invalidProgram(false),
-            _stateCache(State::DEFAULT),
-            _stateSignatureCache(0),
-            _state(nullptr),
-            _invalidState(false)
-             {}
+        GLState* _state { nullptr };
+        bool _invalidState { false };
     } _pipeline;
-
-    // Output stage
-    void do_setFramebuffer(Batch& batch, size_t paramOffset);
-    void do_clearFramebuffer(Batch& batch, size_t paramOffset);
-    void do_blit(Batch& batch, size_t paramOffset);
-    void do_generateTextureMips(Batch& batch, size_t paramOffset);
 
     // Synchronize the state cache of this Backend with the actual real state of the GL Context
     void syncOutputStateCache();
     void resetOutputStage();
     
     struct OutputStageState {
-
-        FramebufferPointer _framebuffer = nullptr;
-        GLuint _drawFBO = 0;
-        
-        OutputStageState() {}
+        FramebufferPointer _framebuffer { nullptr };
+        GLuint _drawFBO { 0 };
     } _output;
-
-    // Query section
-    void do_beginQuery(Batch& batch, size_t paramOffset);
-    void do_endQuery(Batch& batch, size_t paramOffset);
-    void do_getQuery(Batch& batch, size_t paramOffset);
 
     void resetQueryStage();
     struct QueryStageState {
         
     };
 
-    // Reset stages
-    void do_resetStages(Batch& batch, size_t paramOffset);
-
-    void do_runLambda(Batch& batch, size_t paramOffset);
-
-    void do_startNamedCall(Batch& batch, size_t paramOffset);
-    void do_stopNamedCall(Batch& batch, size_t paramOffset);
-
     void resetStages();
-
-    // TODO: As long as we have gl calls explicitely issued from interface
-    // code, we need to be able to record and batch these calls. THe long 
-    // term strategy is to get rid of any GL calls in favor of the HIFI GPU API
-    void do_glActiveBindTexture(Batch& batch, size_t paramOffset);
-
-    void do_glUniform1i(Batch& batch, size_t paramOffset);
-    void do_glUniform1f(Batch& batch, size_t paramOffset);
-    void do_glUniform2f(Batch& batch, size_t paramOffset);
-    void do_glUniform3f(Batch& batch, size_t paramOffset);
-    void do_glUniform4f(Batch& batch, size_t paramOffset);
-    void do_glUniform3fv(Batch& batch, size_t paramOffset);
-    void do_glUniform4fv(Batch& batch, size_t paramOffset);
-    void do_glUniform4iv(Batch& batch, size_t paramOffset);
-    void do_glUniformMatrix4fv(Batch& batch, size_t paramOffset);
-
-    void do_glColor4f(Batch& batch, size_t paramOffset);
-
-    void do_pushProfileRange(Batch& batch, size_t paramOffset);
-    void do_popProfileRange(Batch& batch, size_t paramOffset);
-    
-    int _currentDraw { -1 };
 
     typedef void (GLBackend::*CommandCall)(Batch&, size_t);
     static CommandCall _commandCalls[Batch::NUM_COMMANDS];
-
+    friend class GLState;
 };
 
 } }
-
-Q_DECLARE_LOGGING_CATEGORY(gpugllogging)
-
 
 #endif
