+++ libraries/gpu-gl/src/gpu/gl/GLBackendOutput.cpp
@@ -9,179 +9,13 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 #include "GLBackend.h"
+#include "GLShared.h"
+#include "GLFramebuffer.h"
 
 #include <QtGui/QImage>
 
-#include "GLBackendShared.h"
-
 using namespace gpu;
 using namespace gpu::gl;
-
-GLBackend::GLFramebuffer::GLFramebuffer() {}
-
-GLBackend::GLFramebuffer::~GLFramebuffer() {
-    if (_fbo != 0) {
-        glDeleteFramebuffers(1, &_fbo);
-    }
-}
-
-GLBackend::GLFramebuffer* GLBackend::syncGPUObject(const Framebuffer& framebuffer) {
-    GLFramebuffer* object = Backend::getGPUObject<GLBackend::GLFramebuffer>(framebuffer);
-
-    bool needsUpate { false };
-    if (!object || 
-        framebuffer.getDepthStamp() != object->_depthStamp ||
-        framebuffer.getColorStamps() != object->_colorStamps) {
-        needsUpate = true;
-    }
-
-    // If GPU object already created and in sync
-    if (!needsUpate) {
-        return object;
-    } else if (framebuffer.isEmpty()) {
-        // NO framebuffer definition yet so let's avoid thinking
-        return nullptr;
-    }
-
-    // need to have a gpu object?
-    if (!object) {
-        // All is green, assign the gpuobject to the Framebuffer
-        object = new GLFramebuffer();
-        Backend::setGPUObject(framebuffer, object);
-        glGenFramebuffers(1, &object->_fbo);
-        (void)CHECK_GL_ERROR();
-    }
-
-    if (needsUpate) {
-        GLint currentFBO = -1;
-        glGetIntegerv(GL_DRAW_FRAMEBUFFER_BINDING, &currentFBO);
-        glBindFramebuffer(GL_FRAMEBUFFER, object->_fbo);
-
-        GLTexture* gltexture = nullptr;
-        TexturePointer surface;
-        if (framebuffer.getColorStamps() != object->_colorStamps) {
-            if (framebuffer.hasColor()) {
-                object->_colorBuffers.clear();
-                static const GLenum colorAttachments[] = {
-                    GL_COLOR_ATTACHMENT0,
-                    GL_COLOR_ATTACHMENT1,
-                    GL_COLOR_ATTACHMENT2,
-                    GL_COLOR_ATTACHMENT3,
-                    GL_COLOR_ATTACHMENT4,
-                    GL_COLOR_ATTACHMENT5,
-                    GL_COLOR_ATTACHMENT6,
-                    GL_COLOR_ATTACHMENT7,
-                    GL_COLOR_ATTACHMENT8,
-                    GL_COLOR_ATTACHMENT9,
-                    GL_COLOR_ATTACHMENT10,
-                    GL_COLOR_ATTACHMENT11,
-                    GL_COLOR_ATTACHMENT12,
-                    GL_COLOR_ATTACHMENT13,
-                    GL_COLOR_ATTACHMENT14,
-                    GL_COLOR_ATTACHMENT15 };
-
-                int unit = 0;
-                for (auto& b : framebuffer.getRenderBuffers()) {
-                    surface = b._texture;
-                    if (surface) {
-                        gltexture = GLBackend::syncGPUObject(surface, false); // Grab the gltexture and don't transfer
-                    } else {
-                        gltexture = nullptr;
-                    }
-
-                    if (gltexture) {
-                        glFramebufferTexture2D(GL_FRAMEBUFFER, colorAttachments[unit], GL_TEXTURE_2D, gltexture->_texture, 0);
-                        object->_colorBuffers.push_back(colorAttachments[unit]);
-                    } else {
-                        glFramebufferTexture2D(GL_FRAMEBUFFER, colorAttachments[unit], GL_TEXTURE_2D, 0, 0);
-                    }
-                    unit++;
-                }
-            }
-            object->_colorStamps = framebuffer.getColorStamps();
-        }
-
-        GLenum attachement = GL_DEPTH_STENCIL_ATTACHMENT;
-        if (!framebuffer.hasStencil()) {
-            attachement = GL_DEPTH_ATTACHMENT;
-        } else if (!framebuffer.hasDepth()) {
-            attachement = GL_STENCIL_ATTACHMENT;
-        }
-
-        if (framebuffer.getDepthStamp() != object->_depthStamp) {
-            auto surface = framebuffer.getDepthStencilBuffer();
-            if (framebuffer.hasDepthStencil() && surface) {
-                gltexture = GLBackend::syncGPUObject(surface, false); // Grab the gltexture and don't transfer
-            }
-    
-            if (gltexture) {
-                glFramebufferTexture2D(GL_FRAMEBUFFER, attachement, GL_TEXTURE_2D, gltexture->_texture, 0);
-            } else {
-                glFramebufferTexture2D(GL_FRAMEBUFFER, attachement, GL_TEXTURE_2D, 0, 0);
-            }
-            object->_depthStamp = framebuffer.getDepthStamp();
-        }
-
-
-        // Last but not least, define where we draw
-        if (!object->_colorBuffers.empty()) {
-            glDrawBuffers((GLsizei)object->_colorBuffers.size(), object->_colorBuffers.data());
-        } else {
-            glDrawBuffer( GL_NONE );
-        }
-
-        // Now check for completness
-        GLenum status = glCheckFramebufferStatus(GL_FRAMEBUFFER);
-
-        // restore the current framebuffer
-        if (currentFBO != -1) {
-            glBindFramebuffer(GL_DRAW_FRAMEBUFFER, currentFBO);
-        }
-
-        bool result = false;
-        switch (status) {
-        case GL_FRAMEBUFFER_COMPLETE :
-            // Success !
-            result = true;
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_ATTACHMENT :
-            qCDebug(gpugllogging) << "GLFramebuffer::syncGPUObject : Framebuffer not valid, GL_FRAMEBUFFER_INCOMPLETE_ATTACHMENT.";
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT :
-            qCDebug(gpugllogging) << "GLFramebuffer::syncGPUObject : Framebuffer not valid, GL_FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT.";
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_DRAW_BUFFER :
-            qCDebug(gpugllogging) << "GLFramebuffer::syncGPUObject : Framebuffer not valid, GL_FRAMEBUFFER_INCOMPLETE_DRAW_BUFFER.";
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_READ_BUFFER :
-            qCDebug(gpugllogging) << "GLFramebuffer::syncGPUObject : Framebuffer not valid, GL_FRAMEBUFFER_INCOMPLETE_READ_BUFFER.";
-            break;
-        case GL_FRAMEBUFFER_UNSUPPORTED :
-            qCDebug(gpugllogging) << "GLFramebuffer::syncGPUObject : Framebuffer not valid, GL_FRAMEBUFFER_UNSUPPORTED.";
-            break;
-        }
-        if (!result && object->_fbo) {
-            glDeleteFramebuffers(1, &object->_fbo);
-            return nullptr;
-        }
-    }
-
-    return object;
-}
-
-
-
-GLuint GLBackend::getFramebufferID(const FramebufferPointer& framebuffer) {
-    if (!framebuffer) {
-        return 0;
-    }
-    GLFramebuffer* object = GLBackend::syncGPUObject(*framebuffer);
-    if (object) {
-        return object->_fbo;
-    } else {
-        return 0;
-    }
-}
 
 void GLBackend::syncOutputStateCache() {
     GLint currentFBO;
@@ -298,44 +132,6 @@ void GLBackend::do_clearFramebuffer(Batch& batch, size_t paramOffset) {
         if (glFramebuffer) {
             glDrawBuffers((GLsizei)glFramebuffer->_colorBuffers.size(), glFramebuffer->_colorBuffers.data());
         }
-    }
-
-    (void) CHECK_GL_ERROR();
-}
-
-void GLBackend::do_blit(Batch& batch, size_t paramOffset) {
-    auto srcframebuffer = batch._framebuffers.get(batch._params[paramOffset]._uint);
-    Vec4i srcvp;
-    for (auto i = 0; i < 4; ++i) {
-        srcvp[i] = batch._params[paramOffset + 1 + i]._int;
-    }
-
-    auto dstframebuffer = batch._framebuffers.get(batch._params[paramOffset + 5]._uint);
-    Vec4i dstvp;
-    for (auto i = 0; i < 4; ++i) {
-        dstvp[i] = batch._params[paramOffset + 6 + i]._int;
-    }
-
-    // Assign dest framebuffer if not bound already
-    auto newDrawFBO = getFramebufferID(dstframebuffer);
-    if (_output._drawFBO != newDrawFBO) {
-        glBindFramebuffer(GL_DRAW_FRAMEBUFFER, newDrawFBO);
-    }
-
-    // always bind the read fbo
-    glBindFramebuffer(GL_READ_FRAMEBUFFER, getFramebufferID(srcframebuffer));
-
-    // Blit!
-    glBlitFramebuffer(srcvp.x, srcvp.y, srcvp.z, srcvp.w, 
-        dstvp.x, dstvp.y, dstvp.z, dstvp.w,
-        GL_COLOR_BUFFER_BIT, GL_LINEAR);
-
-    // Always clean the read fbo to 0
-    glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-
-    // Restore draw fbo if changed
-    if (_output._drawFBO != newDrawFBO) {
-        glBindFramebuffer(GL_DRAW_FRAMEBUFFER, _output._drawFBO);
     }
 
     (void) CHECK_GL_ERROR();
