+++ libraries/gpu-gl/src/gpu/gl/GLBackendTransform.cpp
@@ -9,7 +9,6 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 #include "GLBackend.h"
-#include "GLBackendShared.h"
 
 using namespace gpu;
 using namespace gpu::gl;
@@ -51,26 +50,11 @@ void GLBackend::do_setDepthRangeTransform(Batch& batch, size_t paramOffset) {
     }
 }
 
-void GLBackend::initTransform() {
-    glGenBuffers(1, &_transform._objectBuffer);
-    glGenBuffers(1, &_transform._cameraBuffer);
-    glGenBuffers(1, &_transform._drawCallInfoBuffer);
-#ifndef GPU_SSBO_DRAW_CALL_INFO
-    glGenTextures(1, &_transform._objectBufferTexture);
-#endif
-    size_t cameraSize = sizeof(TransformStageState::CameraBufferElement);
-    while (_transform._cameraUboSize < cameraSize) {
-        _transform._cameraUboSize += _uboAlignment;
-    }
-}
-
 void GLBackend::killTransform() {
     glDeleteBuffers(1, &_transform._objectBuffer);
     glDeleteBuffers(1, &_transform._cameraBuffer);
     glDeleteBuffers(1, &_transform._drawCallInfoBuffer);
-#ifndef GPU_SSBO_DRAW_CALL_INFO
     glDeleteTextures(1, &_transform._objectBufferTexture);
-#endif
 }
 
 void GLBackend::syncTransformStateCache() {
@@ -116,64 +100,6 @@ void GLBackend::TransformStageState::preUpdate(size_t commandIndex, const Stereo
 
     // Flags are clean
     _invalidView = _invalidProj = _invalidViewport = false;
-}
-
-void GLBackend::TransformStageState::transfer(const Batch& batch) const {
-    // FIXME not thread safe
-    static std::vector<uint8_t> bufferData;
-    if (!_cameras.empty()) {
-        bufferData.resize(_cameraUboSize * _cameras.size());
-        for (size_t i = 0; i < _cameras.size(); ++i) {
-            memcpy(bufferData.data() + (_cameraUboSize * i), &_cameras[i], sizeof(CameraBufferElement));
-        }
-        glBindBuffer(GL_UNIFORM_BUFFER, _cameraBuffer);
-        glBufferData(GL_UNIFORM_BUFFER, bufferData.size(), bufferData.data(), GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_UNIFORM_BUFFER, 0);
-    }
-
-    if (!batch._objects.empty()) {
-        auto byteSize = batch._objects.size() * sizeof(Batch::TransformObject);
-        bufferData.resize(byteSize);
-        memcpy(bufferData.data(), batch._objects.data(), byteSize);
-
-#ifdef GPU_SSBO_DRAW_CALL_INFO
-        glBindBuffer(GL_SHADER_STORAGE_BUFFER, _objectBuffer);
-        glBufferData(GL_SHADER_STORAGE_BUFFER, bufferData.size(), bufferData.data(), GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-#else
-        glBindBuffer(GL_TEXTURE_BUFFER, _objectBuffer);
-        glBufferData(GL_TEXTURE_BUFFER, bufferData.size(), bufferData.data(), GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_TEXTURE_BUFFER, 0);
-#endif
-    }
-
-    if (!batch._namedData.empty()) {
-        bufferData.clear();
-        for (auto& data : batch._namedData) {
-            auto currentSize = bufferData.size();
-            auto bytesToCopy = data.second.drawCallInfos.size() * sizeof(Batch::DrawCallInfo);
-            bufferData.resize(currentSize + bytesToCopy);
-            memcpy(bufferData.data() + currentSize, data.second.drawCallInfos.data(), bytesToCopy);
-            _drawCallInfoOffsets[data.first] = (GLvoid*)currentSize;
-        }
-
-        glBindBuffer(GL_ARRAY_BUFFER, _drawCallInfoBuffer);
-        glBufferData(GL_ARRAY_BUFFER, bufferData.size(), bufferData.data(), GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-    }
-
-#ifdef GPU_SSBO_DRAW_CALL_INFO
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, TRANSFORM_OBJECT_SLOT, _objectBuffer);
-#else
-    glActiveTexture(GL_TEXTURE0 + TRANSFORM_OBJECT_SLOT);
-    glBindTexture(GL_TEXTURE_BUFFER, _objectBufferTexture);
-    glTexBuffer(GL_TEXTURE_BUFFER, GL_RGBA32F, _objectBuffer);
-#endif
-
-    CHECK_GL_ERROR();
-
-    // Make sure the current Camera offset is unknown before render Draw
-    _currentCameraOffset = INVALID_OFFSET;
 }
 
 void GLBackend::TransformStageState::update(size_t commandIndex, const StereoState& stereo) const {
