+++ libraries/gpu-gl/src/gpu/gl/GLQuery.cpp
@@ -1,0 +1,12 @@
+//
+//  Created by Bradley Austin Davis on 2016/05/15
+//  Copyright 2013-2016 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#include "GLQuery.h"
+
+using namespace gpu;
+using namespace gpu::gl;
