+++ libraries/gpu-gl/src/gpu/gl/GLTextureTransfer.cpp
@@ -5,20 +5,17 @@
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
-#include "GLBackendTextureTransfer.h"
+#include "GLTextureTransfer.h"
 
 #ifdef THREADED_TEXTURE_TRANSFER
 #include <gl/OffscreenGLCanvas.h>
 #include <gl/QOpenGLContextWrapper.h>
 #endif
 
-
-#include "GLBackendShared.h"
+#include "GLShared.h"
 
 using namespace gpu;
 using namespace gpu::gl;
-
-#include "GLBackend.h"
 
 GLTextureTransferHelper::GLTextureTransferHelper() {
 #ifdef THREADED_TEXTURE_TRANSFER
@@ -45,7 +42,7 @@ GLTextureTransferHelper::~GLTextureTransferHelper() {
 }
 
 void GLTextureTransferHelper::transferTexture(const gpu::TexturePointer& texturePointer) {
-    GLBackend::GLTexture* object = Backend::getGPUObject<GLBackend::GLTexture>(*texturePointer);
+    GLTexture* object = Backend::getGPUObject<GLTexture>(*texturePointer);
     Backend::incrementTextureGPUTransferCount();
 #ifdef THREADED_TEXTURE_TRANSFER
     GLsync fence { 0 };
@@ -53,14 +50,14 @@ void GLTextureTransferHelper::transferTexture(const gpu::TexturePointer& texture
     //glFlush();
 
     TextureTransferPackage package { texturePointer, fence };
-    object->setSyncState(GLBackend::GLTexture::Pending);
+    object->setSyncState(GLSyncState::Pending);
     queueItem(package);
 #else
     object->withPreservedTexture([&] {
         do_transfer(*object);
     });
     object->_contentStamp = texturePointer->getDataStamp();
-    object->setSyncState(GLBackend::GLTexture::Transferred);
+    object->setSyncState(GLSyncState::Transferred);
 #endif
 }
 
@@ -78,7 +75,7 @@ void GLTextureTransferHelper::shutdown() {
 #endif
 }
 
-void GLTextureTransferHelper::do_transfer(GLBackend::GLTexture& texture) {
+void GLTextureTransferHelper::do_transfer(GLTexture& texture) {
     texture.createTexture();
     texture.transfer();
     texture.updateSize();
@@ -99,7 +96,7 @@ bool GLTextureTransferHelper::processQueueItems(const Queue& messages) {
             package.fence = 0;
         }
 
-        GLBackend::GLTexture* object = Backend::getGPUObject<GLBackend::GLTexture>(*texturePointer);
+        GLTexture* object = Backend::getGPUObject<GLTexture>(*texturePointer);
         do_transfer(*object);
         glBindTexture(object->_target, 0);
 
@@ -108,7 +105,7 @@ bool GLTextureTransferHelper::processQueueItems(const Queue& messages) {
         glDeleteSync(writeSync);
 
         object->_contentStamp = texturePointer->getDataStamp();
-        object->setSyncState(GLBackend::GLTexture::Transferred);
+        object->setSyncState(GLSyncState::Transferred);
     }
     return true;
 }
