+++ libraries/gpu-gl/src/gpu/gl/GLTextureTransfer.h
@@ -5,12 +5,16 @@
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
+#ifndef hifi_gpu_gl_GLTextureTransfer_h
+#define hifi_gpu_gl_GLTextureTransfer_h
 
 #include <QtGlobal>
-#include <QSharedPointer>
+#include <QtCore/QSharedPointer>
+
 #include <GenericQueueThread.h>
-#include "GLBackendShared.h"
-#include "GLBackend.h"
+
+#include "GLShared.h"
+#include "GLTexture.h"
 
 #ifdef Q_OS_WIN
 #define THREADED_TEXTURE_TRANSFER
@@ -27,6 +31,7 @@ struct TextureTransferPackage {
 
 class GLTextureTransferHelper : public GenericQueueThread<TextureTransferPackage> {
 public:
+    using Pointer = std::shared_ptr<GLTextureTransferHelper>;
     GLTextureTransferHelper();
     ~GLTextureTransferHelper();
     void transferTexture(const gpu::TexturePointer& texturePointer);
@@ -36,10 +41,12 @@ protected:
     void setup() override;
     void shutdown() override;
     bool processQueueItems(const Queue& messages) override;
-    void do_transfer(GLBackend::GLTexture& texturePointer);
+    void do_transfer(GLTexture& texturePointer);
 
 private:
     QSharedPointer<OffscreenGLCanvas> _canvas;
 };
 
 } }
+
+#endif
