+++ libraries/gpu/CMakeLists.txt
@@ -2,3 +2,5 @@ set(TARGET_NAME gpu)
 AUTOSCRIBE_SHADER_LIB(gpu)
 setup_hifi_library()
 link_hifi_libraries(shared)
+
+target_nsight()
